Altered GetAllFoodItems to use a join and changed decimals to doubles for money properties

+++ RestaurantManagement/FoodItem.cs
@@ -37,6 +37,6 @@ namespace RestaurantManagement
         /// <summary>
         /// The retail price charged for a food item.
         /// </summary>
-        public decimal ItemPrice { get; set; }
+        public double ItemPrice { get; set; }
     }
 }
